FIXED SHIP PLACING ISSUE!
Before, checking to see if the ships spots were one block apart still
didn’t make a ship be placed properly; they could be spaced and still
be confirmed. Added a new function called repeat in definitions that
now solves this problem.

Before, this problem was caused when you placed, for example in an
aircraft carrier, 2 blocks together and 3 blocks together in the same
row/column. The program would go through each point, comparing 1 to 0,
2 to 1, but 3 to 2 didn’t exist, so 3 to 4 was compared. Then, the
program compared 4 to 3. All of this satisfied the requirements set
technically, so the ship was allowed to be confirmed.

Now, with the repeat function, “repeats” can be found and not counted
when comparing ships. For example, in the previous situation, 3 could
be compared to 4, and 4 could be compared to 3, bypassing the checker.
Now, once a pair of numbers are compared, they cannot be compared
again.

Repeat takes the (x,y) parameters of the distance function fed to it.
If the function is true, it stores the parameters in a list called
“badlist”. Next time the distance function comes around, it checks if
the x/y parameters are already in the list. If they are, it disregards
the input, thereby fixing the issue.