Fixed Ship Checker Again, Tidied Up File
With the new repeat function, ship checking still wouldn’t work when
you placed 2 squares and 2 squares, with a gap in the middle, (like
this: xxoxx), and then filled the middle spot.

This is because 1 goes to 0, 2 goes to 3, and then 3 tries going to 2
because it is on the edge and has nowhere else to go. 2 could go to 4,
which is in the middle, but in the line with all of the ‘or’ statements
3 comes first. So, this causes an error.

In the new function, you give a list instead of having lots of or
statements. The function evaluates which ordered pairs work with the
distance formula, and return a 1, and which don’t. If the ordered pairs
work, a number is added to ‘num’ and the OP is added to a list.

After the pairs are evaluated, if num is greater than 1, the statement
is automatically True and nothing is added to badlist. This is because
this means the square is connected to two other squares; there is
nothing that can be wrong with how it is placed, or what is placed next
to it. So, the number isn’t added to badlist.

If num is 1, however, the function checks if the ordered pair that
matched isn’t already in badlist. If it is,
it is false. If it isn’t, it is added to badlist, since if the square
only has one connector there could be a problem with the squares
surrounding it. Then, it returns True.

This solves the problem with the
xxoxx situation, because 2 is connected to 3 and 4. As a result,
nothing is added to badlist. So, 3 and 4 can connect back to two
without any problems, even though 2 technically already connected to
them, satisfying the function.

I also tidied up the file to make it follow PEP 8 slightly better.